# public/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "OCPD Insurance Platform API",
        "description": "REST API dla platformy ubezpieczeń odpowiedzialności cywilnej przewoźnika drogowego (OCPD).\n\n## Funkcjonalności\n- Zarządzanie klientami (przewoźnikami)\n- Zarządzanie polisami OCPD\n- Kalkulator składek z oceną ryzyka\n- Obsługa szkód\n- Statystyki dashboardu",
        "version": "1.0.0",
        "contact": {
            "name": "OCPD Support",
            "email": "support@ocpd-platform.pl"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000/api",
            "description": "Development server"
        }
    ],
...
    "paths": {
...
        "/quotes": {
            "get": {
                "tags": [
                    "Quotes"
                ],
                "summary": "Lista wycen",
                "operationId": "getQuotes",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista wycen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Quote"
                                            }
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Quotes"
                ],
                "summary": "Utwórz nową wycenę",
                "operationId": "createQuote",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuoteInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Wycena utworzona"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Quote": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "clientNIP": {
                        "type": "string"
                    },
                    "requestedSumInsured": {
                        "type": "number"
                    },
                    "requestedScope": {
                        "type": "string"
                    },
                    "calculatedPremium": {
                        "type": "number"
                    },
                    "status": {
                        "type": "string"
                    }
                }
            },
            "QuoteInput": {
                "type": "object",
                "required": [
                    "clientNIP",
                    "requestedSumInsured",
                    "requestedScope"
                ],
                "properties": {
                    "clientNIP": {
                        "type": "string"
                    },
                    "requestedSumInsured": {
                        "type": "number"
                    },
                    "requestedScope": {
                        "type": "string"
                    }
                }
            },
...
        },
        "responses": {
            "BadRequest": {
                "description": "Nieprawidłowe dane wejściowe",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Required field missing"
                                }
                            }
                        }
                    }
                }
            },
...
        }
    }
}